add epub support

+++ instructions/0_preparation.ts
@@ -14,29 +14,7 @@ import { Logger } from "../utils/logger";
 export async function preparation() {
   // #1. check if all config.ts has been set up, if not, prompt the user to set up and exit the process.
   {
-    if (!config.originalPath) {
-      Logger.error("Please set up config.originalPath in config.ts");
-      process.exit(1);
-    }
-    if (!existsSync(config.originalPath)) {
-      Logger.error(
-        `The path ${config.originalPath} does not exist. Please check your config.originalPath in config.ts`,
-      );
-      process.exit(1);
-    }
-    if (!config.outputPath) {
-      Logger.error("Please set up config.outputPath in config.ts");
-      process.exit(1);
-    }
-    if (!config.title) {
-      Logger.error("Please set up config.title in config.ts");
-      process.exit(1);
-    }
-    if (!config.dictionaryPath) {
-      Logger.error("Please set up config.dictionaryPath in config.ts");
-      process.exit(1);
-    }
-    if (!existsSync(config.outputPath)) {
+    if (config.outputPath && !existsSync(config.outputPath)) {
       mkdirSync(config.outputPath, { recursive: true });
     }
     if (!existsSync("./json")) {
@@ -46,7 +24,7 @@ export async function preparation() {
       mkdirSync("./books");
     }
     if (!existsSync("./novel_data.json")) {
-      if (existsSync(config.dictionaryPath)) {
+      if (config.dictionaryPath && existsSync(config.dictionaryPath)) {
         cpSync(config.dictionaryPath, "./novel_data.json");
         Logger.info(`Copied: ${config.dictionaryPath} to novel_data.json`);
       } else {
@@ -61,7 +39,7 @@ export async function preparation() {
     }
   }
   // #2. copy all json from config.outputPath to json folder
-  {
+  if (config.outputPath && existsSync(config.outputPath)) {
     const files = readdirSync(config.outputPath).filter((file) =>
       file.endsWith(".json"),
     );
@@ -75,7 +53,7 @@ export async function preparation() {
     });
   }
   // #3. copy all json from config.originalPath to json folder, only if the json file does not exist in json folder, to avoid overwriting the extracted data.
-  {
+  if(config.originalPath && existsSync(config.originalPath)) {
     const files = readdirSync(config.originalPath).filter((file) =>
       file.endsWith(".json"),
     );
@@ -89,7 +67,7 @@ export async function preparation() {
     });
   }
   // #4. convert all json to html
-  {
+  if (existsSync("./json")) {
     const jsonFiles = readdirSync("./json").filter((file) =>
       file.endsWith(".json"),
     );
